requirements.txt
python-neat
graphviz
pygame